Exception handling for mybatis module

+++ src/main/java/com/pewa/common/AbstractMediaDAO.java
@@ -5,19 +5,29 @@ import com.pewa.MediaModel;
 import com.pewa.PewaType;
 import com.pewa.config.ConfigFactory;
 import com.pewa.dao.MyBatisFactory;
+import com.pewa.exceptions.EmptyMapperException;
+import org.apache.ibatis.exceptions.IbatisException;
+import org.apache.ibatis.exceptions.PersistenceException;
 import org.apache.ibatis.session.ExecutorType;
 import org.apache.ibatis.session.SqlSession;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 
 public abstract class AbstractMediaDAO {
 
+    private static final Logger log = LogManager.getLogger(AbstractMediaDAO.class);
     private static final String ADD_SUCCESS = " item added  : ";
     private static final String UPDATE_SUCCESS = " item updated : ";
     private static final String DELETE_SUCCESS = " item deleted : ";
     private static final String SEARCH_SUCCESS = " element(s) found";
     private static final String NOTHING_FOUND = " no item with this ID found : ";
     private static final String DUPLICATE_ITEM = " item already in database : ";
+    private static final String EMPTY_MAPPER = " mapper list is empty";
+    private static final String INTERNAL_ERROR = " Something unexpected happened, contact developer";
+
     private int rowsAffected;
     private List<MediaModel> output;
     private String returnMessage = "";
@@ -41,17 +51,24 @@ public abstract class AbstractMediaDAO {
     protected Results add(MediaModel media) {
         results = new Results();
         rowsAffected = 0;
-        try (SqlSession session = MyBatisFactory.connectionUser().openSession(ExecutorType.SIMPLE, false)) {
-            System.out.println(getMapperList());
-            for (String mapper : getMapperList()) {
-                rowsAffected += session.insert(ConfigFactory.get(mapper), media);
+        if (!getMapperList().isEmpty()) {
+            try (SqlSession session = MyBatisFactory.connectionUser().openSession(ExecutorType.SIMPLE, false)) {
+                for (String mapper : getMapperList()) {
+                    rowsAffected += session.insert(ConfigFactory.get(mapper), media);
+                }
+                session.commit();
+                results.setRowsAffected(rowsAffected);
+                returnMessage = (rowsAffected != 0)
+                        ? getType() + ADD_SUCCESS + getInfoField()
+                        : getType() + DUPLICATE_ITEM + getInfoField();
+            } catch (IbatisException ex) {
+                log.error(ex.getMessage());
+                returnMessage = INTERNAL_ERROR;
             }
-            session.commit();
-            results.setRowsAffected(rowsAffected);
-            returnMessage = (rowsAffected != 0)
-                    ? getType() + ADD_SUCCESS + getInfoField()
-                    : getType() + DUPLICATE_ITEM + getInfoField();
+        } else {
+            returnMessage = EMPTY_MAPPER;
         }
+
         return results.setReturnMessage(returnMessage);
     }
 
@@ -60,16 +77,22 @@ public abstract class AbstractMediaDAO {
     protected Results delete(Integer id) {
         results = new Results();
         rowsAffected = 0;
-        try (SqlSession session = MyBatisFactory.connectionUser().openSession(ExecutorType.SIMPLE, false)) {
-            for (String mapper : getMapperList()) {
-                rowsAffected += session.delete(ConfigFactory.get(mapper), id);
+        if (!getMapperList().isEmpty()) {
+            try (SqlSession session = MyBatisFactory.connectionUser().openSession(ExecutorType.SIMPLE, false)) {
+                for (String mapper : getMapperList()) {
+                    rowsAffected += session.delete(ConfigFactory.get(mapper), id);
+                }
+                session.commit();
+                results.setRowsAffected(rowsAffected);
+                returnMessage = (rowsAffected != 0)
+                        ? getType() + DELETE_SUCCESS + getInfoField()
+                        : getType() + NOTHING_FOUND + getInfoField();
+            } catch (IbatisException ex) {
+                log.error(ex.getMessage());
+                returnMessage = INTERNAL_ERROR;
             }
-            session.commit();
-            results.setRowsAffected(rowsAffected);
-            returnMessage = (rowsAffected != 0)
-                    ? getType() + DELETE_SUCCESS + getInfoField()
-                    : getType() + NOTHING_FOUND + getInfoField();
-//            log.info(returnMessage);
+        } else {
+            returnMessage = EMPTY_MAPPER;
         }
         return results.setReturnMessage(returnMessage);
     }
@@ -78,15 +101,22 @@ public abstract class AbstractMediaDAO {
     // adds number of deleted rows to exsisting results object
     protected Results delete(Results results) {
         rowsAffected = results.getRowsAffected();
-        try (SqlSession session = MyBatisFactory.connectionUser().openSession(ExecutorType.SIMPLE, false)) {
-            for (String mapper : getMapperList()) {
-                rowsAffected += session.delete(ConfigFactory.get(mapper));
+        if (!getMapperList().isEmpty()) {
+            try (SqlSession session = MyBatisFactory.connectionUser().openSession(ExecutorType.SIMPLE, false)) {
+                for (String mapper : getMapperList()) {
+                    rowsAffected += session.delete(ConfigFactory.get(mapper));
+                }
+                session.commit();
+                results.setRowsAffected(rowsAffected);
+                returnMessage = (rowsAffected != 0)
+                        ? getType() + DELETE_SUCCESS + getInfoField()
+                        : getType() + NOTHING_FOUND + getInfoField();
+            } catch (IbatisException ex) {
+                log.error(ex.getMessage());
+                returnMessage = INTERNAL_ERROR;
             }
-            session.commit();
-            results.setRowsAffected(rowsAffected);
-//            returnMessage = (rowsAffected != 0)
-//                    ? getType() + DELETE_SUCCESS + getInfoField()
-//                    : getType() + NOTHING_FOUND + getInfoField();
+        } else {
+            returnMessage = EMPTY_MAPPER;
         }
         return results;
     }
@@ -95,33 +125,47 @@ public abstract class AbstractMediaDAO {
     protected Results update(MediaModel media) {
         results = new Results();
         rowsAffected = 0;
-        try (SqlSession session = MyBatisFactory.connectionUser().openSession(ExecutorType.SIMPLE, false)) {
-            for (String mapper : getMapperList()) {
-                rowsAffected += session.update(ConfigFactory.get(mapper), media);
+        if (!getMapperList().isEmpty()) {
+            try (SqlSession session = MyBatisFactory.connectionUser().openSession(ExecutorType.SIMPLE, false)) {
+                for (String mapper : getMapperList()) {
+                    rowsAffected += session.update(ConfigFactory.get(mapper), media);
+                }
+                session.commit();
+                results.setRowsAffected(rowsAffected);
+                returnMessage = (rowsAffected != 0)
+                        ? getType() + UPDATE_SUCCESS + getInfoField()
+                        : getType() + NOTHING_FOUND + getInfoField();
+            } catch (IbatisException ex) {
+                log.error(ex.getMessage());
+                returnMessage = INTERNAL_ERROR;
             }
-            session.commit();
-            results.setRowsAffected(rowsAffected);
-            returnMessage = (rowsAffected != 0)
-                    ? getType() + UPDATE_SUCCESS + getInfoField()
-                    : getType() + NOTHING_FOUND + getInfoField();
+        } else {
+            returnMessage = EMPTY_MAPPER;
         }
         return results.setReturnMessage(returnMessage);
     }
 
     protected Results search(String request) {
         results = new Results();
-        try (SqlSession session = MyBatisFactory.connectionUser().openSession(ExecutorType.SIMPLE, false)) {
-            String query = new StringBuilder("%")
-                    .append(request)
-                    .append("%")
-                    .toString();
-            output = session.selectList(ConfigFactory.get(getMapperList().get(0)), query);
-            session.commit();
-            output.forEach(results::setEncounters);
+        if (!getMapperList().isEmpty()) {
+            try (SqlSession session = MyBatisFactory.connectionUser().openSession(ExecutorType.SIMPLE, false)) {
+                String query = new StringBuilder("%")
+                        .append(request)
+                        .append("%")
+                        .toString();
+                output = session.selectList(ConfigFactory.get(getMapperList().get(0)), query);
+                session.commit();
+                output.forEach(results::setEncounters);
 
-            returnMessage = (output.isEmpty())
-                    ? getType() + NOTHING_FOUND
-                    : getType() + output.size() + SEARCH_SUCCESS;
+                returnMessage = (output.isEmpty())
+                        ? getType() + NOTHING_FOUND
+                        : getType() + output.size() + SEARCH_SUCCESS;
+            } catch (IbatisException ex) {
+                log.error(ex.getMessage());
+                returnMessage = INTERNAL_ERROR;
+            }
+        } else {
+            returnMessage = EMPTY_MAPPER;
         }
         return results.setReturnMessage(returnMessage);
     }
@@ -129,26 +173,40 @@ public abstract class AbstractMediaDAO {
 
     protected Results get(Integer mediaId) {
         results = new Results();
-        try (SqlSession session = MyBatisFactory.connectionUser().openSession(ExecutorType.SIMPLE, false)) {
-            output = session.selectList(ConfigFactory.get(getMapperList().get(0)), mediaId);
-            session.commit();
-            output.forEach(results::setEncounters);
-            returnMessage = (output.size() == 0)
-                    ? getType() + NOTHING_FOUND + mediaId
-                    : getType() + output.size() + SEARCH_SUCCESS;
+        if (!getMapperList().isEmpty()) {
+            try (SqlSession session = MyBatisFactory.connectionUser().openSession(ExecutorType.SIMPLE, false)) {
+                output = session.selectList(ConfigFactory.get(getMapperList().get(0)), mediaId);
+                session.commit();
+                output.forEach(results::setEncounters);
+                returnMessage = (output.size() == 0)
+                        ? getType() + NOTHING_FOUND + mediaId
+                        : getType() + output.size() + SEARCH_SUCCESS;
+            } catch (IbatisException ex) {
+                log.error(ex.getMessage());
+                returnMessage = INTERNAL_ERROR;
+            }
+        } else {
+            returnMessage = EMPTY_MAPPER;
         }
         return results.setReturnMessage(returnMessage);
     }
 
     protected Results language(String language) {
         results = new Results();
-        try (SqlSession session = MyBatisFactory.connectionUser().openSession(ExecutorType.SIMPLE, false)) {
-            output = session.selectList(ConfigFactory.get(getMapperList().get(0)), language);
-            session.commit();
-            output.forEach(results::setEncounters);
-            returnMessage = (output.size() == 0)
-                    ? getType() + NOTHING_FOUND
-                    : getType() + output.size() + SEARCH_SUCCESS;
+        if (!getMapperList().isEmpty()) {
+            try (SqlSession session = MyBatisFactory.connectionUser().openSession(ExecutorType.SIMPLE, false)) {
+                output = session.selectList(ConfigFactory.get(getMapperList().get(0)), language);
+                session.commit();
+                output.forEach(results::setEncounters);
+                returnMessage = (output.size() == 0)
+                        ? getType() + NOTHING_FOUND
+                        : getType() + output.size() + SEARCH_SUCCESS;
+            } catch (IbatisException ex) {
+                log.error(ex.getMessage());
+                returnMessage = INTERNAL_ERROR;
+            }
+        } else {
+            returnMessage = EMPTY_MAPPER;
         }
         return results.setReturnMessage(returnMessage);
     }
@@ -156,7 +214,6 @@ public abstract class AbstractMediaDAO {
     public abstract List<String> getMapperList();
 
     public abstract String getInfoField();
-
 
 
 }
